fix: 가격 데이터 타입 일관성 개선
- StockPrice 엔티티의 가격 필드를 Integer에서 Float으로 변경
- StockPriceRequestDTO/ResponseDTO의 가격 필드도 Float으로 통일
- 다른 가격 엔티티들(IndexPrice, CalcStockPrice 등)과 일관성 확보

+++ src/main/java/com/chan/stock_portfolio_backtest_api/dto/response/StockPriceResponseDTO.java
@@ -14,10 +14,10 @@ import java.time.LocalDate;
 @Builder
 public class StockPriceResponseDTO {
     private Integer id;
-    private Integer closePrice;
-    private Integer openPrice;
-    private Integer lowPrice;
-    private Integer highPrice;
+    private Float closePrice;
+    private Float openPrice;
+    private Float lowPrice;
+    private Float highPrice;
     private Integer tradeQuantity;
     private Long tradeAmount;
     private Long issuedCount;
